+++ src/main/kotlin/domain/InputValidator.kt
@@ -4,10 +4,11 @@ object InputValidator {
 
     private const val NUMERIC_ERROR_MSG = "[ERROR] 시도 횟수는 숫자 입력으로만 받습니다."
 
-    fun validateIsNumeric(numberOfTry: String) {
+    fun validateIsNumeric(numberOfTry: String): Int {
         require(numberOfTry.isNumeric()) {
             NUMERIC_ERROR_MSG
         }
+        return numberOfTry.toInt()
     }
 
     private fun String.isNumeric(): Boolean {
